move `init` to `from_iterator`
